DELETE/ UserController.java 중복 파일 제거, controller/user/UserController 위치에 있는 파일 사용
