--- src/main/java/houseway/houseway/controller/UserController.java
@@ -1,39 +1,0 @@
-package houseway.houseway.controller;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-import javax.servlet.http.HttpSession;
-
-@Slf4j  // lombok - logging객체 자동 생성
-@Controller
-@RequestMapping("/user") /* 이렇게 안쓴다면 아래에서 적용하면 됨 */
-@RequiredArgsConstructor
-public class UserController {
-
-@GetMapping("/search")
-    public String find() {
-
-    return "views/user/search";
-}
-@GetMapping("/estateDetail")
-    public String estateDetail() {
-
-    return "views/user/estateDetail";
-}
-@GetMapping("/reservation")
-    public String reservOk() {
-
-    return "views/user/reservation";
-}
-
-@GetMapping("/index_demo")
-    public String indexDemo() {
-    return "views/user/index_demo";
-}
-}
